friend_requests/serializers.py
from rest_framework import serializers
from .models import FriendRequest
from users.serializers import UserSerializer

class FriendRequestSerializer(serializers.ModelSerializer):
    sender = UserSerializer(read_only=True)
    receiver = UserSerializer(read_only=True)

    class Meta:
        model = FriendRequest
        fields = ['id', 'sender', 'receiver', 'status', 'created_at']
