requirements.txt
django==5.1
djangorestframework==3.15.2
django-filter==24.3
psycopg2==2.9.9
python-dotenv==1.0.1